Fix crash and hang issues with Pester CodeLens feature
This change fixes a couple of issues with the new Pester CodeLens
feature.  One is that the "Describe" function is not recognized when it
is not cased exactly like that and the other is that the language server
has an internal error that causes it to hang when typing Describe blocks
into a script.

Resolves PowerShell/vscode-powershell#850
Resolves PowerShell/vscode-powershell#851
Resolves PowerShell/vscode-powershell#852