Server should crash on unhandled exceptions in message loop
This change causes all unhandled exceptions to crash the server so that
the user is not left with an inoperative language service.  This was the
expected behavior before the recent redesign changed how the exception
gets propagated all the way back to the EditorServicesHost.  The host
will now throw the unhandled exception from the WaitForCompletion
method causing the process to exit.