Add Diagnostic logging level for protocol message JSON
This change adds a new "Diagnostic" logging level which is used for
logging the JSON contents of protocol messages that are read or written.
This will provide two benefits:

1. Increase the privacy of user logs that are transmitted over GitHub
2. Increase the performance of writing Verbose logs

We still log the high-level message details like message type, method,
and ID so that we can keep track of the message sequence.