Use raw console bgcolor by default in IHostOutput.WriteOutput
This change resolves PowerShell/vscode-powershell#637 which reports that
on Linux and macOS the input prompt is printed out with a black
background, overriding the console's default background color.  The
fix is to use a System.ConsoleColor of -1 as the default rather than
System.ConsoleColor.Black so that .NET's console library will not
override the console's background color.