Fix cluttered debug history
This change fixes an issue where background commands would be added
to PowerShell's history.  This was due to a limitation in the
PowerShell debugger where the only way commands are excluded is if
the command is part of a hard coded set of commands. As a workaround,
a command from that set (prompt) is added as the first statement.

Resolves PowerShell/vscode-powershell#873