Fix #570 Get-Help -ShowWindow errors in PSIC
This fix overrides the built-in PSHost.PrivateData property, which returns null.  The override instead returns an empty PSObject.  This fixes the Get-Help -ShowWindow scenario since Windows PowerShell will no longer get a NullRefEx on access the host PrivateData property without doing a null check.  Since PrivateData["Window"] will still return null, PowerShell will show the dialog with the desktop as the parent window. IOW this PR does not fix the issue with the help window not appearing on top.  It simply fixes the error you get in PSIC.