Fix PSES crash that happens if you format an empty PS doc
Invoke-Formatter -ScriptDefinition '' - throws a terminating error:
ParameterBinderValidationException.

I suppose we could have added another catch statement to the InvokePowerShell method
in the AnalysisService but in general, we probably want to know about such bugs.

Also, we could also update the extension to not send a format request for an empty doc
but I'm not sure it's worth the additional code change given corner case scenario.
But I could be convinced otherwise.