Fix error w/Start-EditorServices transcript logging using temp console (#691)
This error happens because when log level is set to Diagnostic, we use
Start-Transcript to log the output of Start-EditorServices.ps1.
But the log filename we we're using was always Start-EditorServices.log.
When we start a temp console debug session, is using the
Start-EditorServices.ps1 script to start the debug session.
That attempts to log to the original log file which is locked.
This causes an error. The fix is to append the log filename
which will be unique per launch of the Start-ES script.