Reduce allocations when parsing files
The previous implementation used LINQ on a relatively hot path,
returned a lazy IEnumerable that was always eagerly converted to a
List or Array and allocated two strings per line when Windows
linebreaks were used.  Profiling indicates that PSES spends a huge
percentage of its time in GC, so switching to a slightly more complex
implementation that allocates less than half as much seems justified.