Add support for a "Show Documentation" quick fix menu entry (#789)
* Add support for a "Show Documentation" quick fix menu entry

This requires a corresponding PR to VSCode to handle the new command:
PowerShell.ShowCodeActionDocumentation

Changed the use of the Diagnostic.Code field to hold the PSSA rule name.
According to the LSP docs, this field is meant to be displayed and
tslint uses it to display the vioated rule. Update the logic to use a
combination of diagnostic fields to get a unique id in order to
retrieve the stored "correction edits" when the
textDocument/codeAction message is processed with the diagnostic
array returned earlier.

Also, observed a few times during startup that RunScriptDiagnostics was
getting called with an empty filesToAnalyze array.  Modified to return
early in that case.

Remove a private overload of DelayThenInvokeDiagnostics that wasn't
being used.

* Address PR comments by switching to StringBuilder

Also update codeAction processing to ensure we don't have multiple
show documentation commands for the same rule.

* Remove extra filesToAnalyze empty array check per PR feedback

* Address PR feedback on StringBuilder

* Fix indentation issue as per PR feedback