Adds ability to use separate pipes for reading and writing (#785)
* Adds ability to use separate pipes for reading and writing

Some clients (e.g. Java on Windows) have blocking File I/O API, thus it is not possible for them to use single named pipe for reading and writing asynchronously. It will add the ability to use separate pipes for reading and writing when the client specifies the "-SplitInOutPipes" switch for the startup script.

* Use Path.DirectorySeparatorChar as read/write pipe name separator for transport config endpoint

* WritePipeServer: pipe direction should be "Out"

* Organize startup script parameters by parameter sets

* EditorServiceTransportType: Introduce separate properties for read/write pipes

* [ PowerShell/PowerShellEditorServices#785 PR ] Cleanup: refactorings, code style and reformat

* remove e in excpe