Fix folding with spanning regions (#805)
* Fix folding for regions with same end token

Previously the token matching was broken;
```
foreach ($1 in $2) {     <----- STARTS MATCH HERE (1)

    $x = @{  <-----   STARTS MATCH HERE (2)
        'abc' = 'def'
    }        <----- ENDS   MATCH HERE (1) (2)
}
```

This was caused by two or more different token pairs sharing the same end token.
This commit modifies the token pair matching to take an array of Start Tokens
instead of a single.  This has the added benefit of performance increase too.

This commit also adds tests for this scenario.

* Add support for braced variable names in folder

Previously the code folder was not aware of braced variable name
assignments.  This commit adds detection and an appropriate test.

* Add negative folding test for splatting

This commit adds a splat command to the folding scenario to ensure that the
folder is not confused by the (at) character for splatting.