Fix unable to open files in problems/peek windows issue
This is due to PSES not properly storing a language client path in the
ClientFilePath property of ScriptFile.  This change ensures that a
ClientFilePath is always stored in the text document Uri that a LSP
client expects.  This code has mostly been provided by
@seeminglyScience. Thanks for the reference to your code.
This fixes issue 1732 in the vscode-powershell repo.