Fix bad context state after debugging (#1210)
Adds a lock around session state eventing. Whenever the lock is held by PowerShell executing a command, the debugger will not try to send events. But whenever the debugger aborting holds the lock, the command processor will wait. This is a workaround to a deeper problem in the threadsafety of the current session state and eventing model and should be fixed later.