fix #1393: Always use local help to return cmdlet help text
If the user has overridden the Get-Help -Online parameter to always be
true, then calls to Get-Help will always launch a new browser window
to the cmdlet help. Instead, we want it to return the full cmdlet text
as intended, otherwise things like intellisense and tooltips may end
up opening the cmdlet help in a new browser window.