Always sign ADO builds
We only run builds on ADO with the intent to release, therefore we need to just
always sign them. Otherwise the compliance scripts' default behavior is to sign
when the build is manually triggered, which is not our workflow. We push a
release branch and expect that automatic build to result in signing.