Set execution policy to `Bypass` when creating the initial runspace
This eliminates the need to have `SetExecutionPolicy` which is a slow function
that first queries a bunch of policies and then within PowerShell itself (as in,
using a cmdlet) sets it to `Bypass`. However, we just want the process scope
(AKA runspace scope) to have the execution policy set to `Bypass` all the time
so that we can import our bundled modules, such as PSReadLine.