Bump OmniSharp to `v0.19.3`
So `v0.19.4` is out on GitHub, but has not yet released to the NuGet
gallery, and I don't wish to delay this any longer. We'll let Dependabot
bring that in as soon as it can be. The only thing missing is the DryIoc
update, which we're simply hoping solves an issue.