Update `modules.json` logic to use only `RequiredVersion`
We want to know exactly which version of each module is being bundled
with the extension, so the min/max logic was removed and replaced with
just a precise version. The field `AllowPrerelease` is now based off the
same parameter given per module because, again, we want to be explicit
about our versions.