Remove local variables from stack frames
Thanks to PowerShell idiosyncracies around scopes, it is impossible for
the debug service to get the "local variables" for each stack frame. The
prior behavior used an assumption that from the 0 index, each increment
to the stack frame corresponded 1-to-1 to an increment in scope, and
that at each of those frames we could get local variables. This does not
work because that 1-to-1 assumption does not hold, as evidenced by
crashes exhibited in the preview extension because there are often more
stack frames than there are scopes. Since this was but a guess, it was
decided that it is worse to provide inaccurate information than a "best
guess" and so each stack frame now only gets auto variables.

In the process of fixing this, the variable detection mechanism was
improved to rely on the raw results of `Get-PSCallStack` when available
locally, and only pass it through a serialization and deserialization
process when necessary for remote debugging.

Finally, the `StackFrameDetails.Create` method had an ancient TODO from
2019 that we decided to drop the skeleton support for, namely passing
the workspace root path and processing the invoation information.

Co-authored-by: Andy Schwartzmeyer <andrew@schwartzmeyer.com>
Co-authored-by: Patrick Meinecke <SeeminglyScience@users.noreply.github.com>