Fix debugger's "auto" variables container
We want this container to hold everything that is most likely
contextually relevant to the user while debugging. It is specifically
NOT analagous to PowerShell's automatic variables, but to Visual
Studio's "auto" view of a debugger's variables. This means that for the
top of the callstack, we merge the local scope's variables with those in
the frame and filter to our own curated list.

We also cleaned up our variable property rehydration path.