Fix regression with `F5` to use `.` instead of `&` operator
This also corrects the scope of the script's parameter variables to be
`Local` instead of `Command` in the `DebuggerAcceptsScriptArgs` test,
and `Local` instead of `Auto` for most of the rest of the tests (as they
were previous to the pipeline work). Excitingly, this even fixes the
`DebuggerSetsVariablesNoConversion` test, although
`DebuggerSetsVariablesWithConversion` is still broken (for a different
reason).