Handle `debuggerResult` being null in `ProcessDebuggerResult`
Not entirely sure why this can be the case, but a stacktrace from a null
dereference crash indicates it can be null. Since the existing logic is
only doing something if the `ResumeAction` field is not null, I assume
we similarly do nothing if the entire object is also null (since that
implies the former, too).