Enforce code style in build and convert errors to suggestions
Instead of using `silence` we now use `suggestion` which enables
Roslynator and other tools to continue to show the warning, but not
block the build. This allows us to set `EnforceCodeStyleInBuild` now,
without having to fix everything, which helps prevent future mistakes.

Note that many of these should be fixed ASAP, and that we previously
thought we'd enabled all the maintainability warnings, but since they
are not turned on by default, the category does nothing.

Co-authored-by: Andy Schwartzmeyer <andrew@schwartzmeyer.com>