Fix running untitled scripts with arguments (but break line breakpoints)
The extant hack that enabled line breakpoints in untitled scripts is
untenable. It shifted the user's args by one since it ran the untitled
script as the first arg to an inline script `. $args[0]` with `$args[0]`
being the script contents, instead of the expected `.  {<script>}`. This
hack that conflicted with the correct passing of arguments to the
debugged script. We are prioritizing the latter over the former, as
command breakpoints and `Wait-Debugger` work as expected, and the extant
behavior was rather unique (and unexpected) in the world of VS Code
extensions, to the point that the extension can now automatically prompt
to save untitled files.