Avoid recording debugger commands in the history
This changes how we decide to run a given command through
`ExecuteInDebugger` or `ExecuteNormally`. It now restricts the former
such that it is only used for PowerShell's intrinsic debugger commands
(like `s`, `c`, etc.) and the latter is used for any other user command
or, more importantly, our own implementation's debugger commands (like
when we call `Get-Variable`).