"Cancel" `Console.ReadKey()` via `sendKeyPress()` notification to client
This "rewrite" uses the `console.sendText()` API in VS Code to respond
to a notification and send a fake key press, allowing us to "cancel"
`ReadKey` by getting it to return. Since we know we requested the
cancellation, we just drop the fake key press. This means we no longer
have to poll on macOS/Linux, and no longer have the ghost "double
key-press" issue on Windows. This fixes a myriad of problems and opens
up a lot of simplification.