Fix extra prompting and manual debugger commands
This requires that we set the REPL to exit correctly. Note that we don't
have a notion of "disconnect" so we removed that logic. We are assuming
the debugger is "active" a little earlier now, that is, also when we've
received an LSP notification to start it (via a launch or attach
handler). Because we receive a notification on "disconnect" even for a
script just finishing, we need to know that the server is not active and
so not cancel the current prompt.