Fix completion text for clients which don’t support `TextEdit`
We could set the `Label` to `CompletionText` as the fallback, but that
then makes the list of completions in fancier clients not as
accurate (since PowerShell explicitly provides `ListItemText` for this
purpose). Instead, we set usually unused field `InsertText` with the
same information we provided for more advanced clients in `TextEdit`.