Fix piping to native commands for Windows PowerShell (#1836)
For an unknown reason, we were explicitly overriding
`Console.InputEncoding`, which led to a bug where input piped to a
native command in Windows PowerShell (where the default encoding is
different than our override) would fail. Instead, we just don't override
the encodings at all, which lets PSES be closer to the expected
experience.

The regression test covers this, though to test the failure I had to
move that override outside its guard (as it wasn't being set during
tests).

Also fix building on ARM64 in Windows PowerShell, which doesn't have
`$IsWindows`.