Fix logic checking for untitled or raw scripts (#1839)
This was much more easily accomplished by checking if what we were given
exists as a file path, in which case we use it as-is (a file path, that
must exist, since we launch it directly). Otherwise we can assume that
it's a raw script block, or a URI from the client that probably
represents an untitled file.