Add `Directory.Exists()` check to `SetInitialWorkingDirectoryAsync()`
There was an edge case where, after setting and un-setting the
`powershell.cwd` property in the VS Code client, instead of being `null`
the setting would be the empty string `""`. The client did not care
about this, and so left the server to silently crash on startup. We
previously assumed that by the time the server started, the directory
would be validated by the client, but this was wrong. So now we check
its existence first.