Remove unnecessary `PowerShellProcessArchitecture` (#1964)
This simply wasn't being used and was overly complicated. The only time
we want the architecture is when queried via LSP so that the VS Code
client can determine which installer to download for its auto-update
feature. If we can, we should deduplicate version logic in the loader
and the `VersionUtils` class.