Guard expansion of `PSPropertyInfo.Value`
Which fixes our test that gets the properties of a process object, since
the value for `ExitCode` fails (as the process hasn't ended), we now
handle that a bit more gracefully and can get all 130 properties.

This also seemed to fix the overall bug where lots of expected
properties failed to show up. While investigating, I also found many
properties duplicated, seemingly due to being retrieved both off the
PSObject and off the .NET object, so now when we do the latter we check
(by name) that it hasn't already been added.