Migrate `FindReferencesVisitor` to `ReferenceTable.ReferenceVisitor`
This required using a `ScriptFile` for the associated test in
`AstOperationsTests`. The coupling of the visitor to the file seems fine
since that's the _point_ of this improvement: a cached sets of
references for each file. We also have to carefully sort our list of
symbol references for things that expect them in order.