Move reference sorting to unit tests
Since the client doesn't actually care which order we return the
references in, we should skip unnecessarily sorting. It appears to apply
its own sorting to the results regardless of their order, and the
specification doesn't say it needs to be sorted.

Also adjust two tests to account for the definitions being in the
reference lists.