Fix disappearing output bug by resetting `TranscribeOnly`
What a bug! In PowerShell 5.1 our artificial pipeline for event handling
(among other possible PowerShell tasks) could cause the output to just
disappear. Turns out a bug in 5.1 with the transcript was being triggered.
Scripts without `Out-Default` would get it appended with
`-TranscribeOnly True` in order for the transcription to happen...but the
`TranscribeOnly` field could erroneously not be set back to `false`! This
would cause all output to just "disappear." It was fixed in PowerShell 7,
but we still have to work around it for PowerShell 5.1 by resetting it.